% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairpub-package.R
\docType{package}
\name{fairpub-package}
\alias{fairpub}
\alias{fairpub-package}
\title{fairpub: How Fair Are You When You Publish/Cite Scientific Works?}
\description{
Provides a user-friendly way to compute the non-profit and academic friendly ratio of the bibliographic reference list before submitting a manuscript for peer review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/fairpub},\url{https://frbcesab.github.io/fairpub}
  \item Report bugs at \url{https://github.com/frbcesab/fairpub/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{rdev.nc@gmail.com} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nicolas Casajus \email{rdev.nc@gmail.com} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]
}

}
\keyword{internal}
