% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_clean_doi.R
\name{fp_clean_doi}
\alias{fp_clean_doi}
\title{Clean a DOI vector}
\usage{
fp_clean_doi(doi)
}
\arguments{
\item{doi}{a \code{character} vector with Digital Object Identifiers (DOI).}
}
\value{
A \code{character} of DOI without prefix and in lower case.
}
\description{
This helper cleans DOIs (Digital Object Identifier) by removing prefix
(\verb{doi:}, \verb{https://doi.org/} and \verb{http://dx.doi.org/}) and using lower case.
}
\examples{
dois <- c(
  "10.1098/rsos.160384",
  "10.1098/RSOS.160384",
  "doi: 10.1098/rsos.160384",
  "http://dx.doi.org/10.1098/rsos.160384",
  "https://doi.org/10.1098/rsos.160384",
  "HTTPS://DOI.ORG/10.1098/RSOS.160384"
)

fp_clean_doi(dois)
}
