% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_extract_doi.R
\name{fp_extract_doi}
\alias{fp_extract_doi}
\title{Extract DOI from a BibTeX file}
\usage{
fp_extract_doi(file)
}
\arguments{
\item{file}{a \code{character} of length 1. The (absolute or relative) path to
the BibTeX file to open.}
}
\value{
A \code{character} vector with extracted DOI. Some values can be \code{NA} in case of
books, chapters, etc. or if references are malformed.
}
\description{
This function reads a BibTex file and extracts the DOI of references (if
originally present in the file).
}
\examples{
# Path to the BibTeX provided by <fairpub> ----
filename <- system.file(
  file.path("extdata", "references.bib"),
  package = "fairpub"
)

# Extract DOI from BibTeX ----
fp_extract_doi(filename)
}
