% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_get_journal_fairness.R
\name{fp_get_journal_fairness}
\alias{fp_get_journal_fairness}
\title{Get the fairness status of a journal}
\usage{
fp_get_journal_fairness(journal)
}
\arguments{
\item{journal}{a \code{character} of length 1. The name of the journal. Do not use
journal abbreviation.}
}
\value{
A \code{data.frame} with two columns: \code{journal}, the journal name, and
\code{fairness}, the fairness status with the following possible values:
\itemize{
\item Non-profit and academic friendly
\item For-profit and academic friendly
\item For-profit and non academic friendly
\item Record not found in OpenAlex
\item Record not found in DAFNEE database
}
}
\description{
By querying the OpenAlex bibliographic database (\url{https://openalex.org}) and
the DAFNEE database (\url{https://dafnee.isem-evolution.fr/}), this function
returns the business model and the academic friendly status of a journal.
}
\examples{
# Be polite and send your email to OpenAlex API ----
options(openalexR.mailto = 'anonymous@mail.com')

\dontrun{
# Fairness status ----
fp_get_journal_fairness("Science")
#>   journal                           fairness
#> 1 Science   Non-profit and academic friendly

# Fuzzy search ----
fp_get_journal_fairness("Science of Nature")
#> No exact match found!
#> The fuzzy search returns these three best candidates:
#>   'The Science of Nature'
#>   'Science Advances'
#>   'People and Nature'

fp_get_journal_fairness("The Science of Nature")
#>                 journal                               fairness
#> 1 The Science of Nature   For-profit and non-academic friendly
}
}
