% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_get_openalex_author_id.R
\name{fp_get_openalex_author_id}
\alias{fp_get_openalex_author_id}
\title{Get OpenAlex author ID}
\usage{
fp_get_openalex_author_id(author = NULL, n = 10)
}
\arguments{
\item{author}{a \code{character} vector of length 1. Name of the author.}

\item{n}{an \code{integer} of length 1. Number of results to return (between
1 and 200, default is \code{10}).}
}
\value{
A data frame with the following columns:
\describe{
\item{id}{OpenAlex author ID}
\item{display_name}{Author name in OpenAlex}
\item{orcid}{ORCID identifier}
\item{works_count}{Number of publications}
}
}
\description{
Queries the OpenAlex bibliographic database (\url{https://openalex.org}) to
retrieve an author's identifier.
}
\examples{
\dontrun{
# Be polite and send your email to OpenAlex API ----
options(openalexR.mailto = 'anonymous@mail.com')

fp_get_openalex_author_id("Nicolas Casajus")
#>            id    display_name               orcid works_count
#> 1 A5004806463 Nicolas Casajus 0000-0002-5537-5294         102

fp_get_openalex_author_id("Nicolas Mouquet")
#>            id    display_name               orcid works_count
#> 1 A5001034207 Nicolas Mouquet 0000-0003-1840-6984         210
}
}
