% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_get_openalex_doi.R
\name{fp_get_openalex_doi}
\alias{fp_get_openalex_doi}
\title{Get OpenAlex publication DOI}
\usage{
fp_get_openalex_doi(title = NULL, n = 10)
}
\arguments{
\item{title}{a \code{character} vector of length 1. The title of the publication.}

\item{n}{an \code{integer} of length 1. Number of results to return (between
1 and 200, default is \code{10}).}
}
\value{
A data frame with the following columns:
\describe{
\item{display_name}{Title of the publication}
\item{publication_year}{Year of publication}
\item{source_display_name}{Journal or source name}
\item{doi}{Digital Object Identifier (DOI)}
}
}
\description{
Queries the OpenAlex bibliographic database (\url{https://openalex.org}) to
retrieve metadata about publications matching a given title, including
their DOI.
}
\examples{
\dontrun{
# Be polite and send your email to OpenAlex API ----
options(openalexR.mailto = 'anonymous@mail.com')

# Search for an full title ----
fp_get_openalex_doi(
  "Citation self-awareness for a fairer academic publishing landscape"
)
#>                                       display_name publication_year
#> 1 Citation self-awareness for a fairer academic...             2026
#>   source_display_name                    doi
#> 1          BioScience 10.1093/biosci/biag028

# Search for a partial title ----
fp_get_openalex_doi("Citation fairer academic landscape")
#>                                       display_name publication_year
#> 1     Strategic citations for a fairer academic...             2025
#> 2 Citation self-awareness for a fairer academic...             2026
#>                       source_display_name                       doi
#> 1 bioRxiv (Cold Spring Harbor Laboratory) 10.1101/2025.08.06.668908
#> 2                              BioScience    10.1093/biosci/biag028
}
}
