% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_list_openalex_work_types.R
\name{fp_list_openalex_work_types}
\alias{fp_list_openalex_work_types}
\title{List valid OpenAlex work types}
\usage{
fp_list_openalex_work_types()
}
\value{
a \code{character} vector of valid OpenAlex work types.
}
\description{
Returns the set of work types recognized by the OpenAlex database and used
for filtering bibliographic records.
}
\details{
These work types correspond to the classification system used by
OpenAlex to describe scholarly outputs. They can be used to filter
results in functions such as \code{\link[=fp_get_openalex_author_works]{fp_get_openalex_author_works()}}.
}
\examples{
fp_list_openalex_work_types()
}
