% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_retrieve_article_doi.R
\name{fp_retrieve_article_doi}
\alias{fp_retrieve_article_doi}
\title{Retrieve the DOI of an article from OpenAlex}
\usage{
fp_retrieve_article_doi(title, n = 5)
}
\arguments{
\item{title}{a \code{character} of length 1. The title of the article.}

\item{n}{an \code{integer} of length 1. The number of results to return (between
1 and 200). Default is \code{5}.}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{display_name}, the publication titles matching the query
\item \code{publication_year}, the year of publication
\item \code{source_display_name}, the name of the journal
\item \code{doi}, the DOI of the publications
}
}
\description{
This function queries the OpenAlex bibliographic database
(\url{https://openalex.org}) to retrieve the DOI of an article based on its
title.
}
\examples{
\dontrun{
# Be polite and send your email to OpenAlex API ----
options(openalexR.mailto = 'anonymous@mail.com')

# Search for an full title ----
fp_retrieve_article_doi(
  "Citation self-awareness for a fairer academic publishing landscape"
)
#>                                       display_name publication_year
#> 1 Citation self-awareness for a fairer academic...             2026
#>   source_display_name                    doi
#> 1          BioScience 10.1093/biosci/biag028

# Search for a partial title ----
fp_retrieve_article_doi(
  "Citation fairer academic landscape"
)
#>                                       display_name publication_year
#> 1     Strategic citations for a fairer academic...             2025
#> 2 Citation self-awareness for a fairer academic...             2026
#>                       source_display_name                       doi
#> 1 bioRxiv (Cold Spring Harbor Laboratory) 10.1101/2025.08.06.668908
#> 2                              BioScience    10.1093/biosci/biag028
}
}
